clean up navigation launch files
