+++ src/localization/launch/gps.launch.py
@@ -3,6 +3,8 @@ import os
 import ament_index_python.packages
 import launch
 import launch_ros.actions
+from launch.actions import IncludeLaunchDescription
+from launch.launch_description_sources import PythonLaunchDescriptionSource
 from ament_index_python.packages import get_package_share_directory
 from launch.substitutions import LaunchConfiguration
 from launch.actions import DeclareLaunchArgument
@@ -22,9 +24,9 @@ def generate_launch_description():
     )
 
     navsat_remappings = [
-        ("imu", "zed/imu_data"),
+        ("imu", "zed/zed_node/imu/data"),
         ("gps/fix", "gps/fix"),
-        ("odometry/filtered", "odometry/filtered/globaousterl"),
+        ("odometry/filtered", "odometry/filtered/global"),
         ("odometry/gps", "gps/odom"),
     ]
 
@@ -39,13 +41,7 @@ def generate_launch_description():
 
     return launch.LaunchDescription(
         [
-            ublox_gps_node,
             navsat_node,
-            launch.actions.RegisterEventHandler(
-                event_handler=launch.event_handlers.OnProcessExit(
-                    target_action=ublox_gps_node,
-                    on_exit=[launch.actions.EmitEvent(event=launch.events.Shutdown())],
-                )
-            ),
+            ublox_cmd,
         ]
     )
