+++ src/localization/launch/rtabmap.launch.py
@@ -39,16 +39,21 @@ def generate_launch_description():
             "publish_tf": False,
             "use_sim_time": use_sim_time,
             "qos_camera_info": 2,
+            "topic_queue_size": 2,
+            "sync_queue_size": 10,
+            "wait_for_transform": 0.3,
             "Grid/DepthDecimation": "2",
             "Grid/RangeMin": "0.5",
-            "Grid/RangeMax": "10.0",
+            "Grid/RangeMax": "5.0",
             "Grid/MinClusterSize": "10",
-            "Grid/MaxGroundAngle": "30",
+            "Grid/MaxGroundAngle": "45",
             "Grid/NormalK": "10",
-            "Grid/CellSize": "0.05",
+            "Grid/CellSize": "0.1",
             "Grid/FlatObstacleDetected": "false",
+            "Grid/MapFrameProjection": "true",
             "GridGlobal/UpdateError": "0.01",
             "GridGlobal/MinSize": "200",
+            "GridGlobal/OccupancyThr": "0.5",
             "Reg/Strategy": "1",
         }
     ]
@@ -85,7 +90,7 @@ def generate_launch_description():
                 ],
             ),
             Node(
-                condition=IfCondition(launch_rtabmapviz_cmd),
+                condition=IfCondition(launch_rtabmapviz),
                 package="rtabmap_viz",
                 executable="rtabmap_viz",
                 output="screen",
