+++ src/navigation/launch/navigation.launch.py
@@ -1,309 +1,68 @@
-# Copyright (C) 2023  Miguel Ángel González Santamarta
-
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-
 import os
 from ament_index_python.packages import get_package_share_directory
 from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument, GroupAction, SetEnvironmentVariable
+from launch.actions import DeclareLaunchArgument, IncludeLaunchDescription
+from launch.launch_description_sources import PythonLaunchDescriptionSource
+from launch.substitutions import LaunchConfiguration
 from launch.conditions import IfCondition
-from launch.substitutions import LaunchConfiguration, PythonExpression
-from launch_ros.actions import LoadComposableNodes
-from launch_ros.actions import Node
-from launch_ros.descriptions import ComposableNode
-from nav2_common.launch import RewrittenYaml
+from pathlib import Path
 
 
 def generate_launch_description():
-    # Get the launch directory
-    bringup_dir = get_package_share_directory("nav2_bringup")
-    pkg_dir = get_package_share_directory("navigation")
 
-    namespace = LaunchConfiguration("namespace")
-    use_sim_time = LaunchConfiguration("use_sim_time")
-    autostart = LaunchConfiguration("autostart")
-    params_file = LaunchConfiguration("params_file")
-    use_composition = LaunchConfiguration("use_composition")
-    container_name = LaunchConfiguration("container_name")
-    use_respawn = LaunchConfiguration("use_respawn")
-    cmd_vel_topic = LaunchConfiguration("cmd_vel_topic")
-    default_bt_xml_filename = LaunchConfiguration("default_nav_to_pose_bt_xml")
+    # Get package directory
+    pkg_navigation = get_package_share_directory("navigation")
 
-    lifecycle_nodes = [
-        "controller_server",
-        "smoother_server",
-        "planner_server",
-        "behavior_server",
-        "bt_navigator",
-        "waypoint_follower",
-        "velocity_smoother",
-    ]
-
-    # Map fully qualified names to relative ones so the node"s namespace can be prepended.
-    # In case of the transforms (tf), currently, there doesn"t seem to be a better alternative
-    # https://github.com/ros/geometry2/issues/32
-    # https://github.com/ros/robot_state_publisher/pull/30
-    # TODO(orduno) Substitute with `PushNodeRemapping`
-    #              https://github.com/ros2/launch_ros/issues/56
-    remappings = [
-        ("/tf", "tf"),
-        ("/tf_static", "tf_static"),
-        ("/cmd_vel", cmd_vel_topic),
-    ]
-
-    # Create our own temporary YAML files that include substitutions
-    param_substitutions = {
-        "use_sim_time": use_sim_time,
-        "default_nav_to_pose_bt_xml": default_bt_xml_filename,
-        "autostart": autostart,
-    }
-
-    configured_params = RewrittenYaml(
-        source_file=params_file,
-        root_key=namespace,
-        param_rewrites=param_substitutions,
-        convert_types=True,
+    # Declare launch arguments
+    use_sim_time = LaunchConfiguration("use_sim_time", default="false")
+    launch_transversability = LaunchConfiguration(
+        "launch_transversability", default="True"
     )
+    launch_zed = LaunchConfiguration("launch_zed", default="True")
 
-    stdout_linebuf_envvar = SetEnvironmentVariable(
-        "RCUTILS_LOGGING_BUFFERED_STREAM", "1"
+    launch_transversability_cmd = DeclareLaunchArgument(
+        "launch_transversability",
+        description="Launch elevation mapping pkg if True",
+        default_value="True",
     )
-
-    declare_namespace_cmd = DeclareLaunchArgument(
-        "namespace", default_value="", description="Top-level namespace"
+    launch_zed_cmd = DeclareLaunchArgument(
+        "launch_zed", description="Launch description if True", default_value="True"
     )
-
-    declare_use_sim_time_cmd = DeclareLaunchArgument(
+    use_sim_time_cmd = DeclareLaunchArgument(
         "use_sim_time",
+        description="Use simulation (Gazebo) clock if True",
         default_value="false",
-        description="Use simulation (Gazebo) clock if true",
     )
 
-    declare_params_file_cmd = DeclareLaunchArgument(
-        "params_file",
-        default_value=os.path.join(pkg_dir, "params", "nav2.yaml"),
-        description="Full path to the ROS2 parameters file to use for all launched nodes",
+    transversability_launch_file_path = (
+        Path(pkg_navigation) / "launch" / "traversability_gridmap.launch.py"
+    )
+    transversability_cmd = IncludeLaunchDescription(
+        PythonLaunchDescriptionSource(str(transversability_launch_file_path)),
+        condition=IfCondition(launch_transversability),
     )
 
-    declare_autostart_cmd = DeclareLaunchArgument(
-        "autostart",
-        default_value="true",
-        description="Automatically startup the nav2 stack",
+    zed_launch_file_path = Path(pkg_navigation) / "launch" / "zed.launch.py"
+    zed_cmd = IncludeLaunchDescription(
+        PythonLaunchDescriptionSource(str(zed_launch_file_path)),
+        condition=IfCondition(launch_zed),
+        launch_arguments={"use_sim_time": use_sim_time}.items(),
     )
 
-    declare_use_composition_cmd = DeclareLaunchArgument(
-        "use_composition",
-        default_value="False",
-        description="Use composed bringup if True",
+    nav2_cmd = IncludeLaunchDescription(
+        PythonLaunchDescriptionSource(
+            os.path.join(pkg_navigation, "launch", "nav2.launch.py")
+        ),
+        launch_arguments={"use_sim_time": use_sim_time}.items(),
     )
 
-    declare_container_name_cmd = DeclareLaunchArgument(
-        "container_name",
-        default_value="nav2_container",
-        description="the name of conatiner that nodes will load in if use composition",
+    return LaunchDescription(
+        [
+            use_sim_time_cmd,
+            launch_transversability_cmd,
+            launch_zed_cmd,
+            transversability_cmd,
+            nav2_cmd,
+            zed_cmd,
+        ]
     )
-
-    declare_use_respawn_cmd = DeclareLaunchArgument(
-        "use_respawn",
-        default_value="False",
-        description="Whether to respawn if a node crashes. Applied when composition is disabled.",
-    )
-
-    cmd_vel_topic_cmd = DeclareLaunchArgument(
-        "cmd_vel_topic",
-        default_value="/drive/cmd_vel",
-        description="cmd_vel topic (for remmaping)",
-    )
-
-    default_bt_xml_filename_cmd = DeclareLaunchArgument(
-        "default_nav_to_pose_bt_xml",
-        default_value=os.path.join(pkg_dir, "behavior_trees", "behaviorTree.xml"),
-        description="Full path to the behavior tree xml file to use",
-    )
-
-    load_nodes = GroupAction(
-        condition=IfCondition(PythonExpression(["not ", use_composition])),
-        actions=[
-            Node(
-                package="nav2_controller",
-                executable="controller_server",
-                output="screen",
-                respawn=use_respawn,
-                respawn_delay=2.0,
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            Node(
-                package="nav2_smoother",
-                executable="smoother_server",
-                name="smoother_server",
-                output="screen",
-                respawn=use_respawn,
-                respawn_delay=2.0,
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            Node(
-                package="nav2_planner",
-                executable="planner_server",
-                name="planner_server",
-                output="screen",
-                respawn=use_respawn,
-                respawn_delay=2.0,
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            Node(
-                package="nav2_behaviors",
-                executable="behavior_server",
-                name="behavior_server",
-                output="screen",
-                respawn=use_respawn,
-                respawn_delay=2.0,
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            Node(
-                package="nav2_bt_navigator",
-                executable="bt_navigator",
-                name="bt_navigator",
-                output="screen",
-                respawn=use_respawn,
-                respawn_delay=2.0,
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            Node(
-                package="nav2_waypoint_follower",
-                executable="waypoint_follower",
-                name="waypoint_follower",
-                output="screen",
-                respawn=use_respawn,
-                respawn_delay=2.0,
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            Node(
-                package="nav2_lifecycle_manager",
-                executable="lifecycle_manager",
-                name="lifecycle_manager_navigation",
-                output="screen",
-                parameters=[
-                    {"use_sim_time": use_sim_time},
-                    {"autostart": autostart},
-                    {"node_names": lifecycle_nodes},
-                ],
-            ),
-            Node(
-                package="nav2_velocity_smoother",
-                executable="velocity_smoother",
-                name="velocity_smoother",
-                output="screen",
-                respawn=use_respawn,
-                respawn_delay=2.0,
-                parameters=[configured_params],
-            ),
-        ],
-    )
-
-    load_composable_nodes = LoadComposableNodes(
-        condition=IfCondition(use_composition),
-        target_container=container_name,
-        composable_node_descriptions=[
-            ComposableNode(
-                package="nav2_controller",
-                plugin="nav2_controller::ControllerServer",
-                name="controller_server",
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            ComposableNode(
-                package="nav2_smoother",
-                plugin="nav2_smoother::SmootherServer",
-                name="smoother_server",
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            ComposableNode(
-                package="nav2_planner",
-                plugin="nav2_planner::PlannerServer",
-                name="planner_server",
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            ComposableNode(
-                package="nav2_behaviors",
-                plugin="behavior_server::BehaviorServer",
-                name="behavior_server",
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            ComposableNode(
-                package="nav2_bt_navigator",
-                plugin="nav2_bt_navigator::BtNavigator",
-                name="bt_navigator",
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            ComposableNode(
-                package="nav2_waypoint_follower",
-                plugin="nav2_waypoint_follower::WaypointFollower",
-                name="waypoint_follower",
-                parameters=[configured_params],
-                remappings=remappings,
-            ),
-            ComposableNode(
-                package="nav2_lifecycle_manager",
-                plugin="nav2_lifecycle_manager::LifecycleManager",
-                name="lifecycle_manager_navigation",
-                parameters=[
-                    {
-                        "use_sim_time": use_sim_time,
-                        "autostart": autostart,
-                        "node_names": lifecycle_nodes,
-                    }
-                ],
-            ),
-            ComposableNode(
-                package="nav2_velocity_smoother",
-                plugin="nav2_velocity_smoother::VelocitySmoother",
-                name="velocity_smoother",
-                parameters=[configured_params],
-            ),
-        ],
-    )
-
-    # Create the launch description and populate
-    ld = LaunchDescription()
-
-    # Set environment variables
-    ld.add_action(stdout_linebuf_envvar)
-
-    # Declare the launch options
-    ld.add_action(declare_namespace_cmd)
-    ld.add_action(declare_use_sim_time_cmd)
-    ld.add_action(declare_params_file_cmd)
-    ld.add_action(declare_autostart_cmd)
-    ld.add_action(declare_use_composition_cmd)
-    ld.add_action(declare_container_name_cmd)
-    ld.add_action(declare_use_respawn_cmd)
-    ld.add_action(cmd_vel_topic_cmd)
-    ld.add_action(default_bt_xml_filename_cmd)
-
-    # Add the actions to launch all of the navigation nodes
-    ld.add_action(load_nodes)
-    ld.add_action(load_composable_nodes)
-
-    return ld
